removed autosizer
